requirements.txt
Django
djangorestframework
rav
django-environ
drf-yasg